refactor: Extract Gen1 settings translation into a domain service
Move the Gen1 wire knowledge (settable-param allowlists, GET-vs-setter renames, section layout) out of RestoreDeviceConfig into the pure domain service gen1_settings_translation, exposing restorable_params and wifi_subresources. The literal wire-param pinning tests move with it and now run against the pure functions with no mocked gateway.
