+++ packages/core/src/core/domain/services/__init__.py
@@ -1,0 +1,1 @@
+"""Pure domain services: no I/O, no gateway or repository dependencies."""
